+++ R/url_update.R
@@ -11,7 +11,7 @@
 #' url_update("my_pkg")
 #' }
 #' @export
-url_update <- function(path = ".", results = url_check(path)) {
+url_update <- function(path = ".", results = url_check(path, fail = FALSE)) {
   can_update <- vlapply(results[["New"]], nzchar)
   to_update <- results[can_update, ]
   for (row in seq_len(NROW(to_update))) {